+++ CHANGELOG.md
@@ -4,9 +4,12 @@ Changelog for NeoFS Contract
 ## [Unreleased]
 
 ### Added
+- Structured containers to Container contract storage (#534)
+- `createV2` and `getInfo` methods to Container contract (#534)
 
 ### Changed
 - Initial GAS distributed to alphabet members during deploy is raised from 300 to 10000 (#529)
+- Container contract no longer creates or reads container protobuf storage items (#XXX)
 
 ### Updated
 - NeoGo dependency to 0.113.0 (#521)